+++ Diary/Infrastructure/Network/Endpoint/Endpoint.swift
@@ -7,29 +7,20 @@
 
 import Foundation
 
-final class Endpoint: Requestable {
+final class Endpoint {
     var baseURL: String
     var path: String
     var method: HttpMethod
     var queryParameters: Encodable?
-    var bodyParameters: Encodable?
-    var headers: [String: String]?
-    var sampleData: Data?
-    
+
     init(baseURL: String = "",
          path: String = "",
          method: HttpMethod = .get,
-         queryParameters: Encodable? = nil,
-         bodyParameters: Encodable? = nil,
-         headers: [String: String]? = [:],
-         sampleData: Data? = nil
+         queryParameters: Encodable? = nil
     ) {
         self.baseURL = baseURL
         self.path = path
         self.method = method
         self.queryParameters = queryParameters
-        self.bodyParameters = bodyParameters
-        self.headers = headers
-        self.sampleData = sampleData
     }
 }
